Addressing comments from Marius #2
