+++ tests/system-tests/diskencryption/internal/helper/parsehelper_test.go
@@ -1,4 +1,4 @@
-package parsehelper
+package helper
 
 import (
 	"reflect"
